feat(docs): better index info
The HNSW page now mentions the max size of an indexed vector.

The Sparse HNSW page now introduces pg_sparse, and the bare info needed
to get it running.

One thing to note, is the fact that info about pg_sparse is only
available through a github readme link. This is potentially frigiale. To
be fair, it is equally fragile as any other link in these docs. However,
a local link can be validated through mintlify, making dead link harder
to leave in. This is something to keep in mind, but no action is
required.