Remove Some `.collect::<Vec<_>>()`ing of search results
Changes `SearchState::search()` to return an opaque `impl Iterator<...>`
to avoid one round of collecting search results into a Vec.

From there, additional collects are removed in the `score_bm25()`
and `snippet()` functions by leaking the necessary object instances into
the proper MemoryContext so that the opaque iterator can be used
directly by the `TableIterator`.

There's no noticable perf improvement here (the execution time is
dominated by visibility checking), but this reduced a large query that
returns 21.5M rows from using a max of 3.6G of ram to 2.8G.

Additionally, cleaned up the visibility checking code a bit such that
locked buffers can be reused if we see consecutive ctids on the same
block.  In practice, this doesn't seem to happen often due to docs being
returned in scored order, but avoiding as much buffer locking and
reading as we can is definitely a win.