fix: #1579 -- Organize the `$PGDATA/pg_search/` directory by database oid
As noted in #1579, it's possible for two databases in the same cluster to
end up using the same OID for a `USING bm25` index, which could lead to all
kinds of unknown problems at runtime.

This addresses that by organizing the data directory first by database OID,
then by index OID.

There's a few touch points in this PR because we need to pass the database OID
to `WriterDirectory::from_index_oid` now.