Cleanup the Visibility Checker
While working on Custom Scan stuff, I releazed that there's a slightly
more efficient way of doing the visibility checks we need to perform.
This uses `ReleaseAndReadBuffer`, which Postgres says one should prefer
as it eliminates the separate Release and Read steps.

We also reduce the time the last Buffer's lock is held to be as short as
possible.

This logic was reduced out of Postgres' `nodeindexscan.c` functions.