feat: #1569 -- better represent ctids as u64
This implements a `pg_sys::ItemPointerData` to u64 conversion (and back)
that's different from pgrx' built-in version.

This version doesn't waste 2 bytes in the middle between the BlockNumber
and the OffsetNumber.  On a local ~12G tantivy index, the storage
footprint went from 12,445,683,151 bytes to 12,296,613,773.  That's
about 150M savings or about 2%.