feat: a GUC for `per_tuple_cost`
This introduces a new GUC (`paradedb.per_tuple_cost`) that describes how
much it costs for us to retrieve a single tuple from a tantivy index in
situations that are **NOT** Index Scans.

In our case, it costs **A LOT** and we'd prefer Postgres only consider
not using an Index Scan if there's literally no other option.  As such,
the default for this new GUC is one hundred million!