fix: explicitly call `IndexWriter::wait_merging_threads()`
We need to wait for an open IndexWriter to finish merging, or it
eventually never will.

We choose to do this in the background whenever our `SearchIndexWriter`
wrapper is dropped.

If this turns out to be too aggressive, we could instead move this to
only happen during VACUUM... which is an intersting idea.