feat: new `NPlusOneMergePolicy` that targets that many segments
Introduce a new merge policy that will merge the smallest segments in
order to keep at least "N" segments, and there could be a "plus one"
segment for leftovers.  The idea is that we can roughly keep a minimum
number of segments while merging together the smallest segments.

We also add two new index `WITH` options, `target_segment_count =
integer` which provides the "N" for the above and a `merge_on_insert = boolean`
which dictates if we perform a merge after INSERT/UPDATE/COPY statements.

CREATE INDEX and (auto)VACUUM will always do a merge, using our new
policy.