fix: Normal Index Scans now work correctly underneath a Parallel Path
It was possible for our Index Scan implementation to get confused about
if it was a **Parallel** Index Scan or not.

Basically, if Postgres planned a regular (non-parallel) Index Scan
underneath a path that was parallel, our Index Scan would think it was
operating in parallel mode too, when it actually wasn't.