feat: improve general CustomScan performance
Seeing a nearly 2x performance improvement in selecting large number of
rows from queries such as:

```sql
SELECT * FROM t WHERE ... @@@ ...; -- selects millions of rows
```

We cleanup the VisibilityChecker to do it the same way Postgres'
`nodeIndexscan.c` does it, which saves a little bit of overhead.

We cleanup our `BufferedChannelCollector` to immediately send blocks of
matched documents (as a `Vec<T>`) over the channel.  We also switch to
using a bounded channel (bound to the # of CPUs) -- this is where the
big performance improvement came from.

And as a drive-by, fix a possible segfault in our
`SupportRequestSimplify` processing.