feat: improve general CustomScan performance (#1938)
We cleanup the VisibilityChecker to do it the same way Postgres' `nodeIndexscan.c` does it, which saves a little bit of overhead.

We cleanup our `BufferedChannelCollector` to immediately send blocks of matched documents (as a `Vec<T>`) over the channel.  We also switch to using a bounded channel (bound to the # of CPUs) -- this is where the big performance improvement came from.

And as a drive-by, fix a possible segfault in our `SupportRequestSimplify` processing.