feat: improve performance of `SELECT count(*) FROM t WHERE ... @@@ ...` queries
This improves the performance of queries that specifically only select
`count(*)` from a single relation.  Testing shows it to be nearly 10x
faster.

We get this performance increate because in this case we can use
Postgres' visibility map and avoid heap lookups entirely for any block
that's know to be "all visible".

This does mean that regular (auto)VACUUMs are important, but that's true
anyways.