chore: remove `opstamp` from `SegmentMetaEntry`
We (nor tantivy) actually use the `opstamp` value, so there's no need to
store it.  We now fabricate a fake value values on our stored segment's
`xmin` and `xmax` values.

This saves 8 bytes for every segment entry and for entries that contain
a deletes file, it saves another 8 bytes -- 16 total.