feat: Optimize writing to be exactly `bm25_max_free_space()` bytes
Writing to block storage would prefer we do it 1 full block at a time,
rather than partiall filling a block.  Rust's `std::io::BufWriter` is
incompatabile with this desire, as it will write everything through if
the new set of bytes doesn't fit in the buffer.

We implement our own `ExactBuffer` that keeps a buffer of example
`bm25_max_free_space()` bytes and will write through only in chunks of
that size, if necessary.

Unrelated, during `ambuild()` and `aminsert()`'s initialization, we go
walk the `TupleDescriptor` and `SearchIndexSchema` **one** to categorize
the fields to be indexed, rather than doing that on every row being
inserted.

Combined, these two changes nearly double the indexing
records-per-second rate.