feat: optimize queries with a `LIMIT` but no `ORDER BY`
If a query has a LIMIT but no ORDER BY, we can run this through our
custom scan's "top n" machinery, but essentially picking a random
document.

We endeavor to *avoid* parallel custom scans here, but we will generate
one if the LIMIT happens to be greater than num_segments cubed -- needed
some kind of heuristic and this seems to roughly be the breaking point
on my computer.  We can always tweak this logic in the future