perf: don't `.collect()` search results anymore
In the general cases of querying a segment (or even all segments) we no
longer go through tantivy's `Collector` APIs and instead map the
underlying `Scorer` as an `Iterator`.

This allows us to completely eliminate the need to first collect all the
`(Score, DocAddress)` matches into `Vec`s.

Additionally, it allows for `SearchIndexReader::estimate_docs()` to
avoid a lot of overhead and just use the `::size_hint()`, which is good
enough for our query planning purposes.