perf: optimize scans
This PR optimizes the varous code paths that iterate over `SearchResults` to eliminate overhead that showed up in the profiler.

As some drive-bys, we also swap out a `Mutex` or two with `OnceLock` and remove from `#[derive(Clone)]` and `Arc` requirements.  The former does help with perf in their use cases and the latter makes the code a little easier to reason about.