perf: bypass `pg_sys::IsTransactionState()`, directly linking to the function (#2123)
This changes our use of `pg_sys::IsTransactionState()` to directly link to the function so that we can bypass the sigsetjmp overhead pgrx introduces.

This function, as written in Postgres, does not call `ereport()` so we don't have to worry about the FFI boundary.

Calling `pg_sys::IsTransactionState()` showed up as about 3% of the 66% of the time pg_search spends reading Buffers during a simple query.

With this, it's no longer in the profile at all.