fix: send the list of segment ids to the parallel scan workers
Ever since we first added support for Parallel Index Scans, we've had a
bug where a parallel scan worker could see a different representation of
the index than the main backend connection first did.  The way a worker
"checked out" a segment to query was based simply on decrementing a
count, and assuming that number was the `SegmentOrdinal`.  This is
incorrect when a parallel scan worker opens the index for itself and
possibly sees a different list of segments, in a different order.

This approach was then inherited by our Parallel Custom Scan support,
spreading this bug like a virus.

This PR fixes this problem by passing the actual list of `SegmentId`s
down to the parallel scan workers (in both Parallel Index Scan and
Parallel Custom Scan situations) so they can explictly query the proper
segment

This also removes the duplicated `SpinLock` code and updates our tantivy
rev to
https://github.com/paradedb/tantivy/commit/75dec2cf9596eea24dd912a81b1dfdf190064d1a,
which allows us to construct a `SegmentId` instance from a raw byte
array.