fix: deadlocks be gone!
We had a few deadlocks related to block storage.

The first was that `LinkedItemList::garbage_collect()` would travel to
the next page and exclusively lock it while it still held an exclusive
lock on the current page.  This could cause a lock inversion between
concurrent backends that were traversing the linked list at the same
time.

The second is that `ambulkdelete()` was simply holding the MERGE_LOCK
too long.  Specifically, holding it while we (now temporarily) lock the
CLEANUP_LOCK page could cause deadlocks in concurrent backends.

And while not exactly a deadlock issue, via discussions with
@rebasedming, we've decided we were holding the CLEANUP_LOCK much longer
than we needed to.  So now we simply acquire and release it at the end
of `ambulkdelete()`.  This creates a barrier that concurrent scans will
queue up against to ensure they see the proper doc deleted statuses,
post ambulkdelete.