fix: concurrency bug related to VACUUM deleting segments it shouldn't
Primarily, this PR fixes a concurrency bug with VACCUM where it could
delete segments it shouldn't.  The reason this could happen is that,
eventually, tantivy reloads the IndexMeta from the Directory and we use
that information to compare with the previous version to see if any
segments had been deleted.

The nature of MVCC is such that the reloading phase could see different
entries b/c of concurrent transactions.

The solution is for our [`MVCCDirectory`] to locally cache the results
of `load_metas()` and always return that, ensuring tantivy always sees a
consistent view of the underlying index structure throughout its
updating/commit process.

This PR also introduces a new UDF called `paradedb.find_ctid(regclass, tid)` that
will find the segment containing the specified ctid.  It also expands
the number of columns returned from `paradedb.index_info()` to indicate
the segment visibility.

We also centralize the logic for "list_managed_files()" into
`LinkedItemList` as it's now used in a few places.