fix: interrupt holdoff around threads
We increment `pg_sys::InterruptHoldoffCount` in our
`ChannelRequstHandler` before we send a closure to a background thread
to wait for it to finish.  We then decrement it when that closure has
returned from the background thread.

This ensures that Postgres will ignore interrupts while we're
essentially blocked on a background tantivy thread doing indexing/commit
work.

Also, `ambulkdelete()` no longer inspects the `pg_sys::InterruptPending`
state.  We rely on `pg_sys::vacuum_delay_point()` to process interrupts,
if it even can.

Because of our open MergeLock, it probably can't.  But that's okay as
far as things go.