feat: Support the new tantivy Directory panic handler
We added support for tantivy's `Directory` to have a panic handler, to
be used when a thread pool worker (from committing and merging) panics.

This implements one on MVCCDirectory (and delegats down to it for
ChannelDirectory) that makes a best effort to report the panic back to
Postgres, rolling back the current transaction, rather than causing rust
to abort the whole process, taking down the entire Postgres cluster with
it.