perf: use `NumericFastFieldsExecMethod` when the scan's target list is empty
If we detect we don't need to return any data from the custom scan
(`select count(*) from t where ... @@@ ...` is a common case) then we
can use our `NumericFastFieldsExecMethod` instead of the
`NormalExecMethod`.  This allows us to avoid potentially going to the
heap to check visibility -- we can rely solely on the visibility map.