feat: improved JOIN support
This adds improved Join support, allowing our (Parallel) Custom Scans to
be used for the left and right sides of the JOIN.

It also ensures that snippets and scores are pulled up from each side of
the join and go to their proper Const pointer slot in the projection
list.

A small handful of tests that already tested scores-from-joins to
reflect the **correct** score (the tests/code were wrong) now that
things are more correctly implemented.

This PR also adds some basic integration points for parameterized plans,
however they are not supported.  We will raise an ERROR if asked to
execute a parameterized plan.  Until they are fully supported, users
should `SET plan_cache_mode = force_custom_plan;` in their session or
directly in `postgresql.conf`.