fix: issue #2187 - properly support recursive executor calls
For Postgres versions before v17 we have to fake a system similar to the
`aminsertcleanup()` function by hooking the executor.

Our efforts to do that were not nearly complete enough and didn't
account for the fact the postgres executor is reentrant.

Now we hook the ExecutorRun_hook along with the ExecutorFinish_hook and
push/pop a stack a of our own state tracking information so we know what
indexes have been modified by `aminsert()` at that level of the
executor.  This allows us to do the cleanup work necessary for that
level.

Closes #2187