fix: issue #2198 -- don't truncate TEXT `key_field`s
Prior to this PR, we would trucate TEXT `key_field` values to 255
characters.  This undoes that, effectively allowing a TEXT key_field
value to be as long as necessary.

This will potentially require users that use TEXT primary keys to
REINDEX their `USING bm25` indexes as now our underding of the index's
schema does not match what's on disk.  This would lead to incorrect
search results due to tokenization differences.