fix: issue #2199 - UUID fields need to use the `raw` tokenizer/normalizer
We index fields of type UUID as text and until now their values were
being tokenized using Tantivy's default tokenizer.  Which in the case of
well-formed UUIDs would split the value into 5 tokens, splitting on the
dash (`-`).

This is problematic because it is then impossible for the (at least)
`paradedb.term()` function to find docs matching a full UUID value.

This PR will necessitate that users that store UUIDs perform a REINDEX
of all their `USING bm25` indexes so that the index schema matches what
we believe to be true.