fix: issue #2202 - smarter merge decisions based on `paradedb.create_index_parallelism`
If the parallelism during an INSERT/UPDATE/COPY was 1 and more than 1
segment was created, then we will **NOT** merge.

Otherwise, we will attempt a merge.  This accounts for a DML using
multiple threads but only inserting a handful of records (relative to a
completely full segment, anyways), ensuring that the individually
"small" segments will get merged into the index, keeping the index's
segment count near the target of "available parallelism".

This also does some plumbing work to pass the used "parallelism" value
down to `MVCCDirectory` and does some minor cleanup around the
`MergeLock` type.

We also add a new administrative UDF called `paradedb.reset_num_segments()`
which will change the on-disk target segment count to the user provided
value and force a merge to that.  If the user provided value is less
than "available parallelism", then we go with "available parallelism" and
inform the user.