fix: no longer use a Snapshot to acquire the MergeLock
We can check the "vacuum horizon" (my words) to decide if the last merge
transaction's effects are visible to the backend trying to obtain a
MergeLock, instead of looking at active Snapshot.

In fact, looking at the active Snapshot is not even possible when
operating inside a Logical Replication Worker -- doing so leads to an
assert failure in Postgres.