chore: cleanup our index schema/metadata handling (#2205)
We had code spread across 3 different places that could, in one way or another, collect the field configurations from an index to ultimately instantiate a tantivy Schema (or return to the user, or whatever).

This cleans all that up and centralizes everything into our `SearchIndexCreateOptions` type.

As much as possible, the same error _message_ text, and the order they're detected, was preserved in order to appease the unit tests, however one was rewritten for clarity and another one was added entirely.

One side effect of having all this code centralized is that we can now support having an index defition that only contains one field -- the key field.  So a test has been added to assert that works, and an existing test that asserted it didn't work was updated.