fix: issue #2202 - rewrite `NPlusOneMergePolicy` to better balance segments by size (#2204)
Our approach to segment merging has been completely overhauled by this PR.

The tl;dr is that we now target having "available parallelism" non-frozen segments, with up to 10% more "leftover" small segments.  A segment is considered frozen, which means it is no longer a candidate for merging, if its estimated byte size (based on the # of *live* docs)  is >= the new `paradedb.max_mergeable_segment_size` setting.

Which brings us to the new `paradedb.max_mergeable_segment_size`.  It has an arbitrary default of 200MB.  It's measured in bytes and any segment that we estimate its size to be >= to this value will **not** be considered for merging.

When calculating segments to merge we first identify all segments whose size (in # of docs) is at least a standard deviation below the mean of num_docs.  These segments are iteratively grouped together into `MergeCandidate`s, up to an estimated byte size that is capped at `paradedb.max_mergeable_segment_size `, at which point another MergeCandidate is initialized.

If there are no segments a stddev below the mean (segments are generally well-balanced), but there are still more than "available parallelism" non-frozen segments, then those extra segments are merged together, smallest-to-largest, based on their live doc count.  When a MergeCandidate's doc count is >= the current mean, another MergeCandidate is initialized.

There's also short-circuiting, where we don't merge anything, if the count of stddev-below-the-mean segments are less than our `min_merge_count`.  This helps to avoid merging a newly created single-record segment with an existing, much larger segment, so long as it keeps us at or below N+1.  The idea is that a small "leftovers" segment will improve single-record atomic INSERT/UPDATE performance.

All of this means that an index can grow, organically, in segment count without us needing to track a `num_segments` property in the `MergeLock` page.  As such, that field is now gone.