fix: issue #2229 - resolve failures between `pg_search` and `pg_ivm`
This fixes an obsecure pg_search error that is raised when updating a
base table and pg_ivm is running its dynamically-generated sql to update
the "materialized view" that contains that base table.

The bug is more generic in that pg_search assumed, when looking for
possible fast fields that they all come from the base relation, which isn't
always true, of course.  So this short-circuits when the Var isn't on
the same level as we know we're inspecting