feat: make tantivy aware of pending Postgres interrupts
This allows Tantivy to be aware of any pending Postgres interrupts
(mainly concered about "^C" user, or administrator, query cancel
commands).

The important bit here is that this, when combined with
https://github.com/paradedb/tantivy/pull/31/commits/cbaff7f180dfa0e5c5c962bc734ef0e47e99290f,
makes tantivy index merges cancellable.

This is incredibly important not only from a UX perspective (users need
the ability to cancel INSERT/UPDATE statements, even if they're "stuck"
in the segment merge phase), but it's also important for being able to
quickly/cleanly stop a running Postgres instance that is maybe merging
segments via a logical replication apply worker, or similar.

Cancelling a merge may not happen immediately, but will happen on the
next big "chunk" of data that tantivy will process.  This is a trade-off
to avoid checking against checking for interrupts on every individual
"item" being merged v/s maintaining merge's general performance.