feat: move page recycling to the foreground
Until now, page recycling only happened during the `amvacuumcleanup`
process.  With this PR it happens immediately after a segment merge is
complete (at the end of our `aminsertcleanup`), which is the point in
which pages have otherwise been freed and can be recycled.

The page reclimation itself happens outside of a lock as the segments
its operating on are known to not be in use by any other backend but
that one.

Additionally, `amvacuumcleanup` becomes a no-op.

There's a few minor internal API function name changes to make things a
little easier to understand.  There's also a handful of new asserts to
validate the invariants this requires -- namely that we never try to
recycle pages we're not allowed (our "fixed block numbers list") and
also that we don't mark as deleted pages that are already.

We also adjust our heuristics around getting a free buffer from the FSM
by asking for one of the size we need for our bm25 pages.  The theory
here is that it'll be be much less likely for it to then return a buffer
we can't actually use.