fix: `NPlusOneMergePolicy` better balances segments by num_docs (#2288)
This adjusts things such that when we have more than N segments, we re-calcuate our understanding of the num_docs "mean" to be that of the segments that we're not going to consider merging.

This allows for uniform grouping of segments by their document count, ensuring the resulting segments stay closer together in size, rather than having a single segment that's potentially significantly larger than the existing largest segment.

It also means that we may not merge down to exactly N segments in the first pass, but that's okay as we also want to do as little rewriting as possible on any given merge.

---

While here, also fix the standard deviation calculation to be that of the sample rather than the population.  In practice I'm not sure this makes much of a difference, but it now matches what Postgres' `stddev()` aggregate function would calculate over the same set of integers, and that seems reasonable.