perf: cache `FileEntry`s in `MvccDirectory`
Tantivy lookups up `FileEntry` instanaces quite often, and it's to our
benefit to just cache all of them when we lookup the first one.

If for some reason the requested entry isn't in the cache we'll
re-populate the cache.  This can happen from a tantivy instance that has
internally added a new segment during commit.