feat: new GUC: `paradedb.segment_merge_scale_factor` (#2293)
This new GUC, `paradedb.segment_merge_scale_factor` is an integer that's multiplied by the host's available parallelism at the time we decide if we're going to attempt a merge.  The default is `1` (one).

If the index segment count is less than this calculated value, then we will skip merging.

Also fixes a bug with `NPlusOneMergePolicy` where it would merge all the segments down to one if they all happened to be the same size.