feat: Snapshots are no longer necessary during insert
When insert/updating tuples, specifically for the "merge" portion of the
insert, we no longer require a Postgres `Snapshot`.

This allows us to simplify the code quite a bit and eliminates a
complexity around `aminsert()` that behaves differently between
community and enterprise (with logical replication).

Essentially, we don't need to use the snapshot to determine if a segment
meta entry (or block!) is recyclable, and we don't need a snapshot to
determine a segment is mergable.