feat: Snapshots are no longer necessary during insert (#2354)
When insert/updating tuples, specifically for the "merge" portion of the insert, we no longer require a Postgres `Snapshot`.

Essentially, we don't need to use the snapshot to determine if a segment meta entry (or block!) is recyclable, and we don't need a snapshot to determine if a segment is mergable.