fix: no need to assign a `TransactionId` if we don't have one
`load_metas()` might need to know the current transaction id to
determine if an entry is mergeable, but if we don't have a transaction
id we also don't need to assign one.

This would only happen during a postgres WAL recovery situation, so
isn't even applicable here on community pg_search, but it's best to be
as correct as we can.