feat: add new `paradedb.force_merge()` function
This adds a new UDF named `paradedb.force_merge()` that when given an
index name and a byte size (either as exact byte number or a pretty
string, ie, '10G') will essentially force a merge of segments up to the
specified size -- effectively creating a new layer.

The intent is that the user would specify a byte size that's
(significantly) larger than the largest layer_size configured on the
index, but this is not required.

We also adjust our decision making around what segments a concurrent
merge-after-insert might do such that we calculate the merge candidates
ahead of time, enabling other concurrent merges to have access to more
segments for merging, which includes the new `force_merge()` function.