Change how we determine if a SegmentMetaEntry is recyclable
We used to use the "vacuum horizon" to be the determining factor in
deciding when a segment can be garbage collected and its underlying
storage recycled.

As it turns out, long-running transactions hold that "vacuum horizon"
longer, causing segments to build up in any index, so long as there's a
long-running transaction in the same database.  And the new
`force_merge()` function could be considered a long-running transaction.
Things like "CREATE INDEX [CONCURRENTLY]" can too.

This changeset changes how we determine when a segment is recycleable to
be a coordinated effort between holding a pin on a buffer in a segment
whenever we have an open MVCCDirectory and determining if that buffer
can be locked.  The logic in `block.rs#377` is pretty straightfoward.

We're also able to now undo the "retry" loop in `ambulkdelete()` and
move the `MergeLock::garbage_collect()` there, and remove it from
`amvacuumcleanup()`.