fixup locking/segment visbility problems
I think we had two problems here.  The first is that the segment merge process itself would move tuples from transactions that are known "not in progress anymore" into a transaction that is.  So I've addressed that by ensuring merged segments (the original segments and the resulting merged segment) maintain their "not in progress anymore" transaction status by assigning them to the pg_sys::FrozenTransactionId.

Secondly, due to the concurrent nature of both merging and now garbage collection, it was possible for a backend reading the list of segments to see an incomplete view of the list if there was a concurrent backend updating the list.  Tantivy expects the reads/writes of "IndexMeta" to be atomic.  So I've addressed that by having writers changing the list take an exclusive lock on a page while writing to the list, and readers take a shared lock on the same page while reading the list.