feat: add new `paradedb.force_merge()` function (#2366)
This adds a new UDF named `paradedb.force_merge()` that when given an index name and a byte size (either as exact byte number or a pretty string, ie, '10G') will essentially force a merge of segments up to the specified size -- effectively creating a new layer.

The intent is that the user would specify a byte size that's (significantly) larger than the largest layer_size configured on the index, but this is not required.

We also adjust our decision making around what segments a concurrent merge-after-insert might do such that we calculate the merge candidates ahead of time, enabling other concurrent merges to have access to more segments for merging, which includes the new `force_merge()` function.

---

Along with the above...

We used to use the "vacuum horizon" to be the determining factor in deciding when a segment can be garbage collected and its underlying storage recycled.

As it turns out, long-running transactions hold that "vacuum horizon" longer cause segments to build up in any index.  This happens because they're not recyclable yet as the long-running transaction, regardless of what it's actually doing, _could_ see those segments. And the new `force_merge()` function could be considered a long-running transaction. Things like `CREATE INDEX [CONCURRENTLY]` can too.

However, we don't actually need to worry about the vacuum horizon.  When we delete a segment, it can be recycled as soon as the deleting transaction is no longer in progress and if we know there's no concurrent readers/writers using any part of that segment.

So we change how we determine when a segment is recycleable to be a coordinated effort between holding a pin on a buffer in a segment whenever we have an open `MVCCDirectory` and determining if that buffer can be locked.  The logic in `block.rs#377` is pretty straightfoward.

We're also able to now remove the "retry" loop in `ambulkdelete()` and move the `MergeLock::garbage_collect()` there, and remove it from `amvacuumcleanup()`.

---

Finally, we had two problems with the above.  The first is that the segment merge process itself would move tuples from transactions that are known "not in progress anymore" into a transaction that is.  That's now addressed by ensuring merged segments (the original segments and the resulting merged segment) maintain their "not in progress anymore" transaction status by assigning them to the pg_sys::FrozenTransactionId.

A minor side effect of this is that the transaction doing the merge can't merge any of the segments it created -- only segments that are from transactions that are no longer in progress.  Tests have been adjusted accordingly.

Secondly, due to the concurrent nature of both merging and now garbage collection, it was possible for a backend reading the list of segments to see an incomplete view of the list if there was a concurrent backend updating the list.  Tantivy expects the reads/writes of "IndexMeta" to be atomic.  And that's addressed that by having writers changing the list take an exclusive lock on a page while writing to the list, and readers take a shared lock on the same page while reading the list.

---

And that exposed one oversight which is that parallel custom scan workers, and parallel index scan workers, can't use our "MvccSatisfies::Snapshot" rules and instead must use the new `MvccSatisfies::ParallelWorker` rules, which is literally "every segment". 

This is because the parallel workers address segments by their `segment_id`, but the specific segments could have been merged and marked as deleted between when the leader started and any given parallel worker started.  Because the leader holds a pin on each segment, they're still there -- it's just that the parallel worker might not be able to see it anymore due to the visibility rules for `MvccSatisfies::Snapshot`.  Having the workers simply see all the segments allows them to address a specific segment directly by its segment_id.