feat: merge `MergeCandidates` serially
In order to reduce memory and disk I/O pressure, we now merge the
calulcated list of `MergeCandidates` one at a time.  This is especially
helpful for the `paradedb.force_merge()` UDF.

We also drop the pins on the merged-away segments as soon as we can,
making them subject to recycling by concurrent backends sooner.

Additionally, `paradedb.force_merge()` reports some bit of progress as
it merges.