feat: merge `MergeCandidates` serially (#2383)
In order to reduce memory and disk I/O pressure, we now merge the calulcated list of `MergeCandidates` one at a time.  This is especially helpful for the `paradedb.force_merge()` UDF.

We also drop the pins on the merged-away segments as soon as we can, making them subject to recycling by concurrent backends sooner.

Additionally, `paradedb.force_merge()` reports some bit of progress as it merges.

Having merged-away segments be able to be gc'd and returned to the FSM by a concurrent backend as soon as they're merged is a nice way to recycle index space a lot sooner.

With that, parallel merges, especially if done via `force_merge()` cause a ton of disk I/O and memory pressure, so this relieves as much of that as we can by doing 1 merge at a time.

Under normal operations it's most likely only 1 `MergeCandidate` will be created anyways, so this isn't necessarily a net-negative on parallelism/performance.