fix: `check_for_interrupts!()` after merging
Segment merging could be cancelled due to a user cancel query request,
such as `^C` in psql.  In that case we'd prefer to report that rather
than raising a generic ERROR saying the merge was cancelled.

Largely this is so the proper SQL error code gets sent to the
interrupted client.