feat: `paradedb.force_merge()` now garbage collects after each merge
This teaches `merge_index_with_policy()` how to garbage collect
immediately after doing a merge, which gets the space freed by the merge
returned to the FSM so that the next merge candidate can potentially use
it.

While doing this, it exposed that in the non-force_merge case, we only
need to garbage collect if a merge happened.  This frees up a bit of CPU
and disk I/O in the normal happy path of inserting rows that don't need
to be merged.