chore: internal API changes to allow copying item lists w/o deserialization
This changes `struct Page[Mut]` and `trait BM25Page` to make a
distinction between deserializing an item at an offet (what was
previously called `read_item()`) and reading the opaque `PgItem`, which
incurrs no deserialization overhead.

We now use the latter when `LinkedItemList` makes an atomic copy of the
list, just to reduce some bit of CPU overhead.  Copying lists happens *a
lot*.