fix: stop leaking pages and make pages recyclable earlier (#2424)
First, internal APIs have been changed such that there's no longer a distinction between "mark_deleted" and "return_to_fsm".  Now, we simply `return_to_fsm` individual `BufferMut` instances or larger collections of them, such as with `LinkedBytesList`.  Internally, the underlying page is automatically marked as deleted.

The impetus for this internal API change is that the new `AtomicGuard::commit()` and `AtomicGuard::drop()` functions were leaking old (or new) pages because they failed to `.mark_deleted()`.  Now, that's just not possible.

Secondly, we rename `BM25Page::recyclable()` to `BM25Page::is_reusable()` and change its implementation to simply see if the page's `xmax` value equals the frozen transaction id.  If it does, it's reusable.  There is no longer a reliance on the "vacuum horizon" for this, which allows pages to be recycled much sooner.

Third, `save_new_metas` was leaking the `LinkedBytesLists` for existing ".deletes" files.  What happens now is we fabricate a fake segment meta entry to only represent the old ".deletes" file, configured in such a way as that it'll be immediately recyclable on the next garbage collect run.

And finally, the `LinkedItemList::list_and_pin()` function has been renamed to `for_each_and_pin()` and has been re-tooled to only pin the segments which pass the "accept" function closure.  The old implementation would pin every segment first, then `load_metas` would drop the pins on the segments that weren't visible to it based on its `MvccSatisfies` rules.  The problem with this is that holding pins on all segments, including those that were already known to be recyclable, means they couldn't be recycled at that moment.  This lead to a situation where the FSM was empty because these segments weren't recycled early, so new segments would end up allocating new space.  Only pinning the segments `load_metas` is interested in allows concurrent garbage collection/recycling to happen eagerly as it's no longer timing dependent.