chore: internal API changes to allow copying item lists w/o deserialization (#2420)
This changes `struct Page[Mut]` and `trait BM25Page` to make a distinction between deserializing an item at an offset (what was previously called `read_item()`) and reading the opaque `PgItem`, which incurs no deserialization overhead.

We now use the latter when `LinkedItemList` makes an atomic copy of the list.