fix: issue #2442 -- using the `normalizer` setting requires `fast:true`
In order to set a normalizer, the field must be a fast field.
Otherwise, while the setting would have no effect, it's very confusing
to know you've set a `normalizer` but then never see it in the
`paradedb.schema()` output.  So raise an error if `normalizer` is set
but the field isn't `fast: true`