fix: Remove unnecessary sorts, and add necessary sorts (#2452)
# Ticket(s) Closed

- Closes #2405

## What

The last piece of #2405 was ensuring that we declare sorted outputs sorted, by marking the resulting plan with the relevant `pathkey`.

But additionally:
* we were occasionally marking parallel-workers results sorted, which is not correct because each worker might consume multiple segments.
* we were not marking some other cases where we push down sorting to Tantivy as resulting in sorted outputs, meaning that Postgres would re-sort them. 

## Why

The early-cutoff feature described in #2405 requires that outputs are declared sorted.

## Tests

Added tests for partitioned early cutoff, and tests that we do not declare any output sorted when parallel workers are used.